Rename eth module files to match convention
- Agreed upon convention for a module is `module/base_module.py`, `module/module.py`, and `module/async_module.py`
