+++ newsfragments/2772.internal.rst
@@ -1,0 +1,1 @@
+Rename the newly-split ``eth`` module files to match convention.
