+++ web3/eth/async_eth.py
@@ -37,7 +37,7 @@ from web3.contract import (
     AsyncContract,
     AsyncContractCaller,
 )
-from web3.eth.base import (
+from web3.eth.base_eth import (
     BaseEth,
 )
 from web3.exceptions import (
